+++ changelog.md
@@ -3,6 +3,7 @@
 ## 14.0.0
 
 * memoizeAsStrings requires Map support
+* chainPromiseNTimes handles rejection like Promise.all
 
 ## 13.1.0